[BugFix] Fixed UTF8 input data.
When reading data from source, cyrillic characters may be displayed as '?????'. This commit fixes this issue. However, it is a requirement that all original input for this library, whether it's comming from a file or form, has to be UTF-8. In the future, the source encoding might become configurable because a lot of UA users might use Windows-1251 by default.